+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - **fixed:** divide by zero panic when calling `map_in_place` with a ZST
 - **added:** allow `map_in_place` to ZST's regardless of alignment
+- **fixed:** many cases where constructing a `FixedBumpVec` of ZSTs resulted in a non-`usize::MAX` capacity
 
 ## 0.10.10 (2024-11-10)
 - **added:** default generic parameters to `BumpPoolGuard`